[FIXES JENKINS-30223] return back com.cloudbees.jenkins.Credential
as of it makes fail to boot jenkins after installation of plugin which depends on this class.
Also remove migration to another package for this class and use it directly
