+++ src/main/java/com/cloudbees/jenkins/GitHubPushTrigger.java
@@ -18,7 +18,6 @@ import jenkins.model.Jenkins.MasterComputer;
 import org.apache.commons.jelly.XMLOutput;
 import org.jenkinsci.plugins.github.GitHubPlugin;
 import org.jenkinsci.plugins.github.config.GitHubPluginConfig;
-import org.jenkinsci.plugins.github.deprecated.Credential;
 import org.jenkinsci.plugins.github.internal.GHPluginConfigException;
 import org.jenkinsci.plugins.github.migration.Migrator;
 import org.kohsuke.stapler.DataBoundConstructor;
